GPII-2468: Construct test nexusWebSocketBoundComponent with createOnEvent
And use an IoCSS selector to register the onWebsocketConnected event
listener rather than a direct reference to "{client}". Referencing the
client component directly in this case will fail because the event
'binder' sequence member must be resolvable before the preceding
'executor' sequence member is run (which in our case it is not as the
preceding 'executor' member fires an event to create the client). We can
use a direct reference to "{client}" in future 'binder' sequence members
as the client component will have been created by the time we reach
those later positions in the sequence.